requirements.txt
streamlit
pillow
pytesseract
deep-translator
pyttsx3